Move peggy stuff to actions
