--- ui/core/src/api/PeggyService/index.ts
@@ -1,1 +1,0 @@
-export * from "./PeggyService";
